+++ tests/unit/test_mpic_orchestrator.py
@@ -1,9 +1,9 @@
 import json
-
 import pytest
 import os
 
 import aws_lambda_python.lambda_controller.mpic_orchestrator_lambda_function as lambda_function
+from aws_lambda_python.lambda_controller.mpic_orchestrator import MpicOrchestrator
 
 
 # noinspection PyMethodMayBeStatic
@@ -29,41 +29,34 @@ class TestMpicOrchestrator:
     def random_select_perspectives_considering_rir_should_select_diverse_rirs_given_list_where_some_share_same_rir(
             self, set_env_variables, requested_perspective_count, expected_unique_rirs):
         perspectives = os.getenv("perspective_names").split("|")  # same split logic as in actual calling code
-        mpic_orchestrator = lambda_function.MpicOrchestrator()
+        mpic_orchestrator = MpicOrchestrator()
         selected_perspectives = mpic_orchestrator.random_select_perspectives_considering_rir(perspectives, requested_perspective_count, "test_identifier")
         assert len(set(map(lambda p: p.split('.')[0], selected_perspectives))) == expected_unique_rirs  # expect 3 unique rirs from setup data
 
     def random_select_perspectives_considering_rir_should_throw_error_given_requested_count_exceeds_total_perspectives(
             self, set_env_variables):
         perspectives = os.getenv("perspective_names").split("|")
-        mpic_orchestrator = lambda_function.MpicOrchestrator()
+        mpic_orchestrator = MpicOrchestrator()
         with pytest.raises(ValueError):
             mpic_orchestrator.random_select_perspectives_considering_rir(perspectives, 10, "test_identifier")  # expect error
 
     def orchestrate_mpic_should_return_error_given_failed_api_version_check(self, set_env_variables):
-        payload = {
+        body = {
             "api-version": "0.0.0",  # invalid version
             "system-params": {"identifier": "test", "perspective-count": 3},
             "caa-details": {"caa-domains": ["example.com"]}
         }
-        # convert payload to json string
-        payload_as_string = json.dumps(payload)
-
-        event = {"path": "test_path", "body": payload_as_string}
+        event = {"path": "test_path", "body": json.dumps(body)}
         result = lambda_function.lambda_handler(event, None)
         assert result["statusCode"] == 400
         assert "api-version-mismatch" in result["body"]
 
     def orchestrate_mpic_should_return_error_given_perspectives_and_perspective_count_both_specified(self, set_env_variables):
-        payload = {
+        body = {
             "api-version": "1.0.0",
-            "system-params": {"identifier": "test", "perspective-count": 3, "perspectives": "test1|test2|test3"},
-            "caa-details": {"caa-domains": ["example.com"]}
+            "system-params": {"identifier": "test", "perspective-count": 3, "perspectives": "test1|test2|test3"}
         }
-        # convert payload to json string
-        payload_as_string = json.dumps(payload)
-
-        event = {"path": "test_path", "body": payload_as_string}
+        event = {"path": "test_path", "body": json.dumps(body)}
         result = lambda_function.lambda_handler(event, None)
         assert result["statusCode"] == 400
         assert "perspectives-and-perspective-count" in result["body"]
@@ -71,15 +64,11 @@ class TestMpicOrchestrator:
     # FIXME: This test is expected to fail because the code does not check for invalid perspective count
     @pytest.mark.xfail
     def orchestrate_mpic_should_return_error_given_invalid_perspective_count(self, set_env_variables):
-        payload = {
+        body = {
             "api-version": "1.0.0",
-            "system-params": {"identifier": "test", "perspective-count": 0},
-            "caa-details": {"caa-domains": ["example.com"]}
+            "system-params": {"identifier": "test", "perspective-count": 0}
         }
-        # convert payload to json string
-        payload_as_string = json.dumps(payload)
-
-        event = {"path": "test_path", "body": payload_as_string}
+        event = {"path": "test_path", "body": json.dumps(body)}
         result = lambda_function.lambda_handler(event, None)
         assert result["statusCode"] == 400
         assert "invalid-perspective-count" in result["body"]
@@ -87,52 +76,23 @@ class TestMpicOrchestrator:
     # FIXME: This test is expected to fail because the code does not check for invalid perspective list
     @pytest.mark.xfail
     def orchestrate_mpic_should_return_error_given_invalid_perspective_list(self, set_env_variables):
-        payload = {
+        body = {
             "api-version": "1.0.0",
-            "system-params": {"identifier": "test", "perspectives": "test1|test2|test3|test4"},
-            "caa-details": {"caa-domains": ["example.com"]}
+            "system-params": {"identifier": "test", "perspectives": "test1|test2|test3|test4"}
         }
-        # convert payload to json string
-        payload_as_string = json.dumps(payload)
-
-        event = {"path": "test_path", "body": payload_as_string}
+        event = {"path": "test_path", "body": json.dumps(body)}
         result = lambda_function.lambda_handler(event, None)
         assert result["statusCode"] == 400
         assert "invalid-perspectives" in result["body"]
 
-    # FIXME: This test is expected to fail because the code does not dynamically calculate required quorum size
-    @pytest.mark.xfail
-    @pytest.mark.parametrize("requested_perspective_count, expected_quorum_size", [(4, 3), (5, 4), (6, 4)])
-    def orchestrate_mpic_should_dynamically_set_required_quorum_count_given_no_quorum_specified(
-            self, set_env_variables,requested_perspective_count, expected_quorum_size):
-        payload = {
-            "api-version": "1.0.0",
-            "system-params": {"identifier": "test", "perspective-count": requested_perspective_count},
-            "caa-details": {"caa-domains": ["example.com"]}
-        }
-        # convert payload to json string
-        payload_as_string = json.dumps(payload)
-
-        event = {"path": "test_path", "body": payload_as_string}
-        result = lambda_function.lambda_handler(event, None)
-        assert result["statusCode"] == 200
-        assert "required-quorum-count" in result["body"]
-        body_as_dict = json.loads(result["body"])
-        actual_quorum_count = body_as_dict["required-quorum-count"]
-        assert actual_quorum_count == expected_quorum_size
-
     # FIXME: This test is expected to fail because the code does not validate quorum count
     @pytest.mark.xfail
     def orchestrate_mpic_should_return_error_given_quorum_count_too_high(self, set_env_variables):
-        payload = {
+        body = {
             "api-version": "1.0.0",
-            "system-params": {"identifier": "test", "perspective-count": 3, "quorum": 4},
-            "caa-details": {"caa-domains": ["example.com"]}
+            "system-params": {"identifier": "test", "perspective-count": 3, "quorum": 4}
         }
-        # convert payload to json string
-        payload_as_string = json.dumps(payload)
-
-        event = {"path": "test_path", "body": payload_as_string}
+        event = {"path": "test_path", "body": json.dumps(body)}
         result = lambda_function.lambda_handler(event, None)
         assert result["statusCode"] == 400
         assert "invalid-quorum-count" in result["body"]
@@ -140,34 +100,58 @@ class TestMpicOrchestrator:
     # FIXME: This test is expected to fail because the code does not validate quorum count
     @pytest.mark.xfail
     def orchestrate_mpic_should_return_error_given_quorum_count_too_low(self, set_env_variables):
-        payload = {
+        body = {
             "api-version": "1.0.0",
-            "system-params": {"identifier": "test", "perspective-count": 3, "quorum": 1},
-            "caa-details": {"caa-domains": ["example.com"]}
+            "system-params": {"identifier": "test", "perspective-count": 3, "quorum": 1}
         }
-        # convert payload to json string
-        payload_as_string = json.dumps(payload)
-
-        event = {"path": "test_path", "body": payload_as_string}
+        event = {"path": "test_path", "body": json.dumps(body)}
         result = lambda_function.lambda_handler(event, None)
         assert result["statusCode"] == 400
         assert "invalid-quorum-count" in result["body"]
 
-    # FIXME: This test is expected to fail; if identifier is missing what should happen?
+    # FIXME: This test is expected to fail because the code does not dynamically calculate required quorum size
+    @pytest.mark.xfail
+    @pytest.mark.parametrize("requested_perspective_count, expected_quorum_size", [(4, 3), (5, 4), (6, 4)])
+    def orchestrate_mpic_should_dynamically_set_required_quorum_count_given_no_quorum_specified(
+            self, set_env_variables, requested_perspective_count, expected_quorum_size):
+        body = {
+            "api-version": "1.0.0",
+            "system-params": {"identifier": "test", "perspective-count": requested_perspective_count},
+        }
+        event = {"path": "test_path", "body": json.dumps(body)}
+        result = lambda_function.lambda_handler(event, None)
+        assert result["statusCode"] == 200
+        assert "required-quorum-count" in result["body"]
+        body_as_dict = json.loads(result["body"])
+        actual_quorum_count = body_as_dict["required-quorum-count"]
+        assert actual_quorum_count == expected_quorum_size
+
+    # FIXME: This test is expected to fail; if identifier is missing what should happen? (ditto for all other fields)
     @pytest.mark.xfail
     def orchestrate_mpic_should_return_error_given_missing_identifier(self, set_env_variables):
-        payload = {
+        body = {
             "api-version": "1.0.0",
-            "system-params": {"perspective-count": 3},
-            "caa-details": {"caa-domains": ["example.com"]}
+            "system-params": {"perspective-count": 3}
         }
-        # convert payload to json string
-        payload_as_string = json.dumps(payload)
-
-        event = {"path": "test_path", "body": payload_as_string}
+        event = {"path": "test_path", "body": json.dumps(body)}
         result = lambda_function.lambda_handler(event, None)
         assert result["statusCode"] == 400
         assert "missing-identifier" in result["body"]
+
+    def collect_async_calls_to_issue_should_create_list_of_only_caa_calls_given_caa_check_request_path(self, set_env_variables):
+        body = {
+            "api-version": "1.0.0",
+            "system-params": {"identifier": "test", "perspective-count": 3},
+            "caa-details": {"caa-domains": ["example.com"]}
+        }
+        perspectives_to_use = os.getenv("perspective_names").split("|")
+        # convert payload to json string
+        mpic_orchestrator = MpicOrchestrator()
+        call_list = mpic_orchestrator.collect_async_calls_to_issue("/caa-check", body, perspectives_to_use)
+        # get length of call_list and assert it's 6
+        assert len(call_list) == 6
+        # map each result to its first field (0 index), then convert that map to a set to force uniqueness,then assert that set only contains 'caa'
+        assert set(map(lambda result: result[0], call_list)) == {"caa"}
 
 
 if __name__ == '__main__':
